wrap network construction
